Splitting ConfigCacheLoader and ConfigCacheSaver
