+++ src/Factory/DependencyFactory.php
@@ -24,7 +24,7 @@ class DependencyFactory
 	{
 		$basePath = BasePath::create($directory, $_SERVER);
 		$mode = new App\Mode($basePath);
-		$configLoader = new Cache\ConfigCacheManager($basePath, $mode);
+		$configLoader = new Cache\ConfigCacheLoader($basePath, $mode);
 		$configCache = Factory\ConfigFactory::createCache($configLoader);
 		$profiler = Factory\ProfilerFactory::create($configCache);
 		Factory\DBFactory::init($configCache, $profiler, $_SERVER);
