+++ tests/src/Core/Config/Cache/ConfigCacheLoaderTest.php
@@ -4,13 +4,13 @@ namespace Friendica\Test\Core\Config\Cache;
 
 use Friendica\App;
 use Friendica\Core\Config\Cache\ConfigCache;
-use Friendica\Core\Config\Cache\ConfigCacheManager;
+use Friendica\Core\Config\Cache\ConfigCacheLoader;
 use Friendica\Test\MockedTest;
 use Friendica\Test\Util\VFSTrait;
 use Mockery\MockInterface;
 use org\bovigo\vfs\vfsStream;
 
-class ConfigCacheManagerTest extends MockedTest
+class ConfigCacheLoaderTest extends MockedTest
 {
 	use VFSTrait;
 
@@ -34,10 +34,10 @@ class ConfigCacheManagerTest extends MockedTest
 	 */
 	public function testLoadConfigFiles()
 	{
-		$configCacheManager = new ConfigCacheManager($this->root->url(), $this->mode);
+		$configCacheLoader = new ConfigCacheLoader($this->root->url(), $this->mode);
 		$configCache = new ConfigCache();
 
-		$configCacheManager->loadConfigFiles($configCache);
+		$configCacheLoader->loadConfigFiles($configCache);
 
 		$this->assertEquals($this->root->url(), $configCache->get('system', 'basepath'));
 	}
@@ -55,10 +55,10 @@ class ConfigCacheManagerTest extends MockedTest
 			->at($this->root->getChild('config'))
 			->setContent('<?php return true;');
 
-		$configCacheManager = new ConfigCacheManager($this->root->url(), $this->mode);
+		$configCacheLoader = new ConfigCacheLoader($this->root->url(), $this->mode);
 		$configCache = new ConfigCache();
 
-		$configCacheManager->loadConfigFiles($configCache);
+		$configCacheLoader->loadConfigFiles($configCache);
 	}
 
 	/**
@@ -80,10 +80,10 @@ class ConfigCacheManagerTest extends MockedTest
 			->at($this->root->getChild('config'))
 			->setContent(file_get_contents($file));
 
-		$configCacheManager = new ConfigCacheManager($this->root->url(), $this->mode);
+		$configCacheLoader = new ConfigCacheLoader($this->root->url(), $this->mode);
 		$configCache = new ConfigCache();
 
-		$configCacheManager->loadConfigFiles($configCache);
+		$configCacheLoader->loadConfigFiles($configCache);
 
 		$this->assertEquals('testhost', $configCache->get('database', 'hostname'));
 		$this->assertEquals('testuser', $configCache->get('database', 'username'));
@@ -113,10 +113,10 @@ class ConfigCacheManagerTest extends MockedTest
 			->at($this->root->getChild('config'))
 			->setContent(file_get_contents($file));
 
-		$configCacheManager = new ConfigCacheManager($this->root->url(), $this->mode);
+		$configCacheLoader = new ConfigCacheLoader($this->root->url(), $this->mode);
 		$configCache = new ConfigCache();
 
-		$configCacheManager->loadConfigFiles($configCache);
+		$configCacheLoader->loadConfigFiles($configCache);
 
 		$this->assertEquals('testhost', $configCache->get('database', 'hostname'));
 		$this->assertEquals('testuser', $configCache->get('database', 'username'));
@@ -145,10 +145,10 @@ class ConfigCacheManagerTest extends MockedTest
 			->at($this->root)
 			->setContent(file_get_contents($file));
 
-		$configCacheManager = new ConfigCacheManager($this->root->url(), $this->mode);
+		$configCacheLoader = new ConfigCacheLoader($this->root->url(), $this->mode);
 		$configCache = new ConfigCache();
 
-		$configCacheManager->loadConfigFiles($configCache);
+		$configCacheLoader->loadConfigFiles($configCache);
 
 		$this->assertEquals('testhost', $configCache->get('database', 'hostname'));
 		$this->assertEquals('testuser', $configCache->get('database', 'username'));
@@ -184,9 +184,9 @@ class ConfigCacheManagerTest extends MockedTest
 			->at($this->root->getChild('addon')->getChild('test')->getChild('config'))
 			->setContent(file_get_contents($file));
 
-		$configCacheManager = new ConfigCacheManager($this->root->url(), $this->mode);
+		$configCacheLoader = new ConfigCacheLoader($this->root->url(), $this->mode);
 
-		$conf = $configCacheManager->loadAddonConfig('test');
+		$conf = $configCacheLoader->loadAddonConfig('test');
 
 		$this->assertEquals('testhost', $conf['database']['hostname']);
 		$this->assertEquals('testuser', $conf['database']['username']);
@@ -194,110 +194,5 @@ class ConfigCacheManagerTest extends MockedTest
 		$this->assertEquals('testdb', $conf['database']['database']);
 
 		$this->assertEquals('admin@test.it', $conf['config']['admin_email']);
-	}
-
-	/**
-	 * Test the saveToConfigFile() method with a local.config.php file
-	 */
-	public function testSaveToConfigFileLocal()
-	{
-		$this->delConfigFile('local.config.php');
-
-		$file = dirname(__DIR__) . DIRECTORY_SEPARATOR .
-			'..' . DIRECTORY_SEPARATOR .
-			'..' . DIRECTORY_SEPARATOR .
-			'..' . DIRECTORY_SEPARATOR .
-			'datasets' . DIRECTORY_SEPARATOR .
-			'config' . DIRECTORY_SEPARATOR .
-			'local.config.php';
-
-		vfsStream::newFile('local.config.php')
-			->at($this->root->getChild('config'))
-			->setContent(file_get_contents($file));
-
-		$configCacheManager = new ConfigCacheManager($this->root->url(), $this->mode);
-		$configCache = new ConfigCache();
-
-		$configCacheManager->loadConfigFiles($configCache);
-		$this->assertEquals('admin@test.it', $configCache->get('config', 'admin_email'));
-		$this->assertEquals('!<unset>!', $configCache->get('config', 'test_val'));
-
-		$configCacheManager->saveToConfigFile('config', 'admin_email', 'new@mail.it');
-		$configCacheManager->saveToConfigFile('config', 'test_val', 'Testing$!"$with@all.we can!');
-
-		$newConfigCache = new ConfigCache();
-		$configCacheManager->loadConfigFiles($newConfigCache);
-		$this->assertEquals('new@mail.it', $newConfigCache->get('config', 'admin_email'));
-		$this->assertEquals('Testing$!"$with@all.we can!', $newConfigCache->get('config', 'test_val'));
-	}
-
-	/**
-	 * Test the saveToConfigFile() method with a local.ini.php file
-	 */
-	public function testSaveToConfigFileINI()
-	{
-		$this->delConfigFile('local.config.php');
-
-		$file = dirname(__DIR__) . DIRECTORY_SEPARATOR .
-			'..' . DIRECTORY_SEPARATOR .
-			'..' . DIRECTORY_SEPARATOR .
-			'..' . DIRECTORY_SEPARATOR .
-			'datasets' . DIRECTORY_SEPARATOR .
-			'config' . DIRECTORY_SEPARATOR .
-			'local.ini.php';
-
-		vfsStream::newFile('local.ini.php')
-			->at($this->root->getChild('config'))
-			->setContent(file_get_contents($file));
-		$configCacheManager = new ConfigCacheManager($this->root->url(), $this->mode);
-		$configCache = new ConfigCache();
-
-		$configCacheManager->loadConfigFiles($configCache);
-		$this->assertEquals('admin@test.it', $configCache->get('config', 'admin_email'));
-		$this->assertEquals('!<unset>!', $configCache->get('config', 'test_val'));
-
-		$configCacheManager->saveToConfigFile('config', 'admin_email', 'new@mail.it');
-		$configCacheManager->saveToConfigFile('config', 'test_val', "Testing@with.all we can");
-
-		$newConfigCache = new ConfigCache();
-		$configCacheManager->loadConfigFiles($newConfigCache);
-		$this->assertEquals('new@mail.it', $newConfigCache->get('config', 'admin_email'));
-		$this->assertEquals("Testing@with.all we can", $newConfigCache->get('config', 'test_val'));
-	}
-
-	/**
-	 * Test the saveToConfigFile() method with a .htconfig.php file
-	 * @todo fix it after 2019.03 merge to develop
-	 */
-	public function testSaveToConfigFileHtconfig()
-	{
-		$this->markTestSkipped('Needs 2019.03 merge to develop first');
-		$this->delConfigFile('local.config.php');
-
-		$file = dirname(__DIR__) . DIRECTORY_SEPARATOR .
-			'..' . DIRECTORY_SEPARATOR .
-			'..' . DIRECTORY_SEPARATOR .
-			'..' . DIRECTORY_SEPARATOR .
-			'datasets' . DIRECTORY_SEPARATOR .
-			'config' . DIRECTORY_SEPARATOR .
-			'.htconfig.test.php';
-
-		vfsStream::newFile('.htconfig.php')
-			->at($this->root)
-			->setContent(file_get_contents($file));
-		$configCacheManager = new ConfigCacheManager($this->root->url(), $this->mode);
-		$configCache = new ConfigCache();
-
-		$configCacheManager->loadConfigFiles($configCache);
-		$this->assertEquals('admin@test.it', $configCache->get('config', 'admin_email'));
-		$this->assertEquals('!<unset>!', $configCache->get('config', 'test_val'));
-
-		$configCacheManager->saveToConfigFile('config', 'admin_email', 'new@mail.it');
-		$configCacheManager->saveToConfigFile('config', 'test_val', 'Testing$!"$with@all.we can!');
-
-		$newConfigCache = new ConfigCache();
-		$configCacheManager->loadConfigFiles($newConfigCache);
-		$this->assertEquals('new@mail.it', $newConfigCache->get('config', 'admin_email'));
-		$this->assertEquals('Testing$!"$with@all.we can!', $newConfigCache->get('config', 'test_val'));
 	}
 }
